Changes to solve a bug related to the reading of the parameters file.

+++ include/Photon.h
@@ -1,20 +1,18 @@
 #ifndef PHOTON_H
 #define PHOTON_H
 
-#include <cstdint>
-
 struct Photon
 {
-    uint64_t id;
-    float x, y, z;
-    uint32_t side;
-    uint64_t previousId;
-    uint64_t nextId;
-    uint64_t surfaceId;
+    double id;
+    double x, y, z;
+    double side;
+    double previousId;
+    double nextId;
+    double surfaceId;
 
     Photon() = default;
 
-    Photon(uint64_t i, float xx, float yy, float zz, uint32_t s, uint64_t prev, uint64_t next, uint64_t surf)
+    Photon(double i, double xx, double yy, double zz, double s, double prev, double next, double surf)
         : id(i), x(xx), y(yy), z(zz), side(s), previousId(prev), nextId(next), surfaceId(surf) {}
 };
